Changed logic for start and end time
Amended code to account for Start time & end time:
If this duration is greater than lunch break, there is a possibility that the driver will go for lunch. When this happens, we deduct lunch break from end time to prevent situation of exceeding 6pm from happening.
If this duration is less than lunch break, there is no way driver will go for lunch (or have a situation exceeding 6pm). Hence we do not deduct lunch break from end time.

This is to account for a solution that takes into consideration of customer's requested delivery time window - specifically Node 3 customer with delivery time 1600 to 1700, only 1 hr timeframe.